Renamed local_io to raster_io, cleaned up, and upgraded raster IO tasks
