+++ io/eolearn/io/__init__.py
@@ -4,7 +4,7 @@ A collection of input and output EOTasks
 
 from .geometry_io import GeopediaVectorImportTask, VectorImportTask
 from .geopedia import AddGeopediaFeatureTask
-from .local_io import ExportToTiffTask, ImportFromTiffTask
+from .raster_io import ExportToTiffTask, ImportFromTiffTask
 from .sentinelhub_process import (
     SentinelHubDemTask,
     SentinelHubEvalscriptTask,
