+++ io/requirements.txt
@@ -2,6 +2,7 @@ boto3
 eo-learn-core
 fiona>=1.8.18
 geopandas>=0.8.1
+affine
 rasterio>=1.2.7
 rtree
 sentinelhub>=3.4.4
